requirements.txt
Django==3.0.3
django-address==0.2.1
django-bootstrap4==1.1.1
django-filter==2.2.0
django-phonenumber-field==4.0.0
phonenumbers==8.12.1
six==1.14.0
